fix: int to float (#53)

+++ twelvelabs/models/classify.py
@@ -25,8 +25,8 @@ class ClassifyDetailedScore(BaseModel):
 
 
 class ClassifyClip(BaseModel):
-    start: int
-    end: int
+    start: float
+    end: float
     score: float
     option: str
     prompt: str
